+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Number settings fields no longer accept garbage input (and show a nice spinner as a bonus)
 - The current selection is now remembered on reload / git update
 - Git clone from SSH (using agent) no longer fails with *unsupported credentials type* error
+- Fixed *create item* button being enabled before permissions have been initially set up (with warning message)
 
 ## v1.1.0 (2017-12-18)